extract admin-specific stuff to `admin.wdio.page.js`
